% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_player.R
\name{mod_add_player_ui}
\alias{mod_add_player_ui}
\alias{mod_add_player_server}
\title{mod_add_player_ui and mod_add_player_server}
\usage{
mod_add_player_ui(id)

mod_add_player_server(input, output, session, con_param)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
