% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_form.R
\name{mod_form_ui}
\alias{mod_form_ui}
\alias{mod_form_server}
\title{mod_form_ui and mod_form_server}
\usage{
mod_form_ui(id, couleur = NULL, contrat = NULL, annonces = NULL,
  bouts = NULL, points = NULL)

mod_form_server(input, output, session, con_param, preneur = NULL,
  appele = NULL, annonce_players = NULL, petit_au_bout_succes = NULL,
  petit_au_bout_sens = NULL, disabled_button, new = TRUE, ID, scores,
  players, active_players = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
