% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scores_display.R
\name{mod_scores_display_ui}
\alias{mod_scores_display_ui}
\alias{mod_scores_display_server}
\title{mod_scores_display_ui and mod_scores_display_server}
\usage{
mod_scores_display_ui(id)

mod_scores_display_server(input, output, session, con_param, scores,
  disabled_button)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
