% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stat.R
\name{mod_stat_ui}
\alias{mod_stat_ui}
\alias{mod_stat_server}
\title{mod_stat_ui and mod_stat_server}
\usage{
mod_stat_ui(id)

mod_stat_server(input, output, session, con_param, scores)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
